test: 기사님 회원가입 통합 테스트 추가
